refactor(core): centralize domain subscriptions
